consumer/requirements.txt
kafka-python==2.2.15
cassandra-driver==3.29.2
elasticsearch==8.17.2
